internal(curves): add curve based on delta
This change creates a curve based on the delta (difference) between the
minimum and maximum values of a collection of curves.

As an example, basing fan speed on, say, water temperature - ambient
temperature will auto adjust the fan curve throughout the day as the
room heats up and cools down, keeping noise levels consistent and
eliminating a use case for fan profiles. It might not be useful for most
users but it is a nice option to have.

internal(curves): add changes from code review

Fixed stylistic changes and updated the README