Fix autoconfigure for weird PWMs
This patch fixes fan autoconfiguration for weird PWMs that have non-1:1 value maps.
In particular, it fixes the construction of the RPM curve data to properly use the
PWM value map computed just before.

This patch also adds a config option and flag to set the delay between setting
a PWM value and observing the fan RPM sensor to support a wide range of
controllers and fans. A long delay can lead to system overheating. The default
is 2 seconds.

This patch also changes a few other delays in the initialization code, namely
the delay between setting a PWM value and observing the PWM value (to detect
controllers that alter/quantize/adjust PWM values). It is hardcoded to 5
milliseconds.