Speeds are now always between fan.MinPwm and fan.MaxPwm
even if it is allowed to stop, in which case the speed value previously
was normalized to 0..MaxPwm instead of MinPwm..MaxPwm.
This makes the behavior of fans that have `neverStop` set and ones that
don't very similar, and thus easier to understand/explain/document.

To still allow fans (that don't have `neverStop` set) to stop,
a 0 value from the curve is translated to PWM 0