Enforce lazy binding so go-nvml works
Some distros seem to default to non-lazy binding ("bind now"), which
means that all imported symbols (functions) are resolved when the
appliation is started, instead of when they're first called.
That breaks go-nvml which loads its wrapped C library (libnvidia-ml.so)
with dlopen() but does *not* use dlsym() to load the functions, but
instead relies on them just being there, like it would be the case
when linking a lib dynamically (instead of loading it dynamically).

Explicitly telling `go build` that lazy binding should be used seems
to make it work even on those distros.
Fixes #400 and hopefully #395 as well.