Tweak conversion from curve values to PWM
when NOT using `UseUnscaledCurveValues`

Curve values 1 and 1% (instead of 0) are converted to MinPwm,
255 and 100% are still converted to MaxPwm.
If you're using NeverStop, values smaller than 1 or 1% are also mapped
to MinPwm, otherwise to 0.

This also (slightly) affects the conversion from percent to speed values
between 0 and 255: When doing a simple scale from percent to [0..255],
1% would be 0.39, and 1 would correspond to 2.55%.
THIS IS NOT THE CASE ANYMORE.
Now 1%-100% are scaled to 1-255 and values smaller than 1% become 0.
This way 0% and 0 mean PWM 0 aka "fan stopped" (unless NeverStop is set)
and both 1% and 1 mean MinPwm (no matter if NeverStop is set or not)