Convert and validate LinearCurve steps when loading the config
and rename the corresponding maps in LinearCurveConfig:
`Steps  map[int]string` is now called InSteps and
`FloatSteps` is again called just `Steps`, like before my changes.

This is made possible by the realization that the name used for parsing
fan2go.yaml can be configured with a `mapstructure:"steps"` tag instead
of relying on the struct member name.